+++ tests/testthat/test_misc.R
@@ -205,6 +205,21 @@ test_that('check_outcome works as expected', {
     check_outcome(factor(1:2), reg_spec)
   )
 
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(NULL, reg_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(tibble::new_tibble(list(), nrow = 10), reg_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    fit(reg_spec, ~ mpg, mtcars)
+  )
+
   class_spec <- logistic_reg()
 
   expect_no_error(
@@ -218,6 +233,21 @@ test_that('check_outcome works as expected', {
   expect_snapshot(
     error = TRUE,
     check_outcome(1:2, class_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(NULL, class_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    check_outcome(tibble::new_tibble(list(), nrow = 10), class_spec)
+  )
+
+  expect_snapshot(
+    error = TRUE,
+    fit(class_spec, ~ mpg, mtcars)
   )
 
   # Fake specification to avoid having to load {censored}