+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Improved errors and documentation related to special terms in formulas. See `?model_formula` to learn more. (#770, #1014)
 
+* Improved errors in cases where the outcome column is mis-specified. (#1003)
+
+
 # parsnip 1.1.1
 
 * Fixed bug where prediction on rank deficient `lm()` models produced `.pred_res` instead of `.pred`. (#985)